+++ magick/attributes.c
@@ -157,7 +157,7 @@ static int GenerateIPTCAttribute(Image *image,const char *key)
   count=sscanf(key,"IPTC:%d:%d",&dataset,&record);
   if (count != 2)
     return(False);
-  for (i=0; i < image->iptc_profile.length; i++)
+  for (i=0; i < (int) image->iptc_profile.length; i++)
   {
     if (image->iptc_profile.info[i] != 0x1c)
       continue;
@@ -176,7 +176,7 @@ static int GenerateIPTCAttribute(Image *image,const char *key)
     LiberateMemory((void **) &attribute);
     break;
   }
-  return(i < image->iptc_profile.length);
+  return(i < (int) image->iptc_profile.length);
 }
 
 static long readLongFromBuffer(char **s,unsigned int *len)
@@ -397,7 +397,7 @@ static int Generate8BIMAttribute(Image *image,const char *key)
       continue;
     count=readByteFromBuffer((char **) &info,&length);
     string=(char *) NULL;
-    if ((count > 0) && (count <= length))
+    if ((count > 0) && (count <= (int) length))
       {
         string=(char *) AcquireMemory(count+MaxTextExtent);
         if (string != (char *) NULL)
